requirements.txt
wget==3.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6

pandas==1.4.3
pyod==0.9.5
tensorflow==2.4.1
tensorflow-estimator==2.4.0
torch==1.10.0a0+3fd9dcf
fast-map==0.0.7
